simplify implementation by moving out bend calculation
